[EC-28] refactor: Tag 인라인 스타일 삭제 및 기능 수정
