+++ client/src/css/index.css
@@ -5,7 +5,7 @@
   --primary-100: #b3e56a;
   --primary-200: #57dd79;
   --primary-300: #e56a6e;
-  --primary-400: #ffd9da;
+  --primary-400: #ffb72e;
   --accent-100: #ffd9da;
   --accent-200: #eef7e6;
   --accent-300: #ffea97;
